+++ src/assets/tour/index.tsx
@@ -1,2 +1,2 @@
-export { default as shape1 } from '../img/shape1.png'
-export { default as product1 } from '../img/product1.jpg'
+export { default as shape1 } from '../tour/shape1.png'
+export { default as product1 } from '../tour/product1.jpg'
